bug fix tazz
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@2409 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Graphics/GRSdk.cpp
@@ -76,7 +76,7 @@ void GRSdk::MergeScene(GRSceneIf* scene0, GRSceneIf* scene1){
 	it1 = find(scenes.begin(), scenes.end(), XCAST(scene1));
 	if(it0 == scenes.end() || it1 == scenes.end())
 		return;
-	for(int i = 0; i < scene1->NChildObject(); i++){
+	for(int i = 0; i < (int)scene1->NChildObject(); i++){
 		scene0->AddChildObject(scene1->GetChildObject(i));
 	}
 	scenes.erase(it1);
